+++ qgreenland/util/metadata.py
@@ -1,80 +1,18 @@
 import datetime as dt
-import tempfile
 from pathlib import Path
-from xml.sax.saxutils import escape
-
-import qgis.core as qgc
 
 from qgreenland.constants.paths import FETCH_DATASETS_DIR
 from qgreenland.models.config.layer import Layer
 from qgreenland.util.layer import datasource_dirname
-from qgreenland.util.template import load_template
 
 
-def add_layer_metadata(map_layer: qgc.QgsMapLayer, layer_cfg: Layer) -> None:
-    """Add layer metadata.
+def build_layer_metadata(layer_cfg: Layer) -> str:
+    """Return layer metadata text.
 
-    Renders a jinja template to a temporary file location as a valid QGIS qmd
-    metadata file. This metadata then gets associated with the `map_layer` using
-    its `loadNamedMetadata` method. This metadata gets written to the project
-    file when the layer is added to the `project`.
+    Includes layer description, dataset description, and citation information.
     """
-    qmd_template = load_template('metadata.jinja')
-
-    # Set the layer's tooltip
-    tooltip = _build_layer_tooltip(layer_cfg)
-    map_layer.setAbstract(tooltip)
-
-    # Render the qmd template.
-    abstract = build_layer_abstract(layer_cfg)
-    layer_extent = map_layer.extent()
-    layer_crs = map_layer.crs()
-
-    if layer_cfg.steps:
-        provenance_list = [escape(step.provenance) for step in layer_cfg.steps]
-    else:
-        provenance_list = []
-
-    rendered_qmd = qmd_template.render(
-        provenance_list=provenance_list,
-        abstract=abstract,
-        title=layer_cfg.title,
-        crs_proj4_str=layer_crs.toProj4(),
-        crs_srsid=layer_crs.srsid(),
-        crs_postgres_srid=layer_crs.postgisSrid(),
-        crs_authid=layer_crs.authid(),
-        crs_description=layer_crs.description(),
-        crs_projection_acronym=layer_crs.projectionAcronym(),
-        crs_ellipsoid_acronym=layer_crs.ellipsoidAcronym(),
-        minx=layer_extent.xMinimum(),
-        miny=layer_extent.yMinimum(),
-        maxx=layer_extent.xMaximum(),
-        maxy=layer_extent.yMaximum(),
-    )
-
-    # Write the rendered tempalte to a temporary file
-    # location. `map_layer.loadNamedMetadata` expects a string URI corresponding
-    # to a file on disk.
-    with tempfile.NamedTemporaryFile('w') as temp_file:
-        temp_file.write(rendered_qmd)
-        temp_file.flush()
-        map_layer.loadNamedMetadata(temp_file.name)
-
-
-def _build_layer_tooltip(layer_cfg: Layer) -> str:
-    """Return a properly escaped layer tooltip text."""
-    tt = _build_layer_description(layer_cfg)
-    tt += (
-        '\n\n'
-        'Open Layer Properties and select the Metadata tab for more information.'
-    )
-    return escape(tt)
-
-
-def build_layer_abstract(layer_cfg: Layer) -> str:
-    """Return a properly escaped layer abstract text."""
     # Include the layer description first.
-    abstract = _build_layer_description(layer_cfg)
+    abstract = build_layer_description(layer_cfg)
 
     # If the layer has a description, separate it from the abstract of the
     # original data source.
@@ -89,10 +27,18 @@ def build_layer_abstract(layer_cfg: Layer) -> str:
     # Add the dataset's citation
     abstract += _build_dataset_citation(layer_cfg)
 
-    return escape(abstract)
+    return abstract
 
 
-def _build_layer_description(layer_cfg: Layer) -> str:
+def write_metadata_file(*, layer_cfg: Layer, filepath: Path) -> None:
+    """Write layer metadata to a text file."""
+    with open(filepath, 'w') as metadata_file:
+        metadata_file.write(
+            build_layer_metadata(layer_cfg),
+        )
+
+
+def build_layer_description(layer_cfg: Layer) -> str:
     """Return a string representing the layer's description."""
     layer_description = ''